#147: Moved PropertyTestRunner to fj.test.runner
